+++ utilities/encodefloat.cpp
@@ -3,12 +3,12 @@ using namespace std;
 
 int main(int argc, char* argv[])
 {
-	size_t g;
+	float g;
 	const auto a = reinterpret_cast<char*>(&g);
 	const auto s = sizeof(g);
 	for (string line; getline(cin, line);)
 	{
-		g = stoul(line);
+		g = stof(line);
 		cout.write(a, s);
 	}
 }
